thanks void

+++ Vent173/Config.cs
@@ -9,6 +9,6 @@ namespace Vent173
         [Description("Enable the plugin?")]
         public bool IsEnabled { get; set; } = true;
         [Description("How long is the vent cooldown?")]
-        public int VentCooldown { get; set; } = 30;
+        public float VentCooldown { get; set; } = 30;
     }
 }
